Move mapping handlers to their own file
Rename handlers.go to differentiate between the two.
